+++ HaremKing/HaremKingFemalePatch.cs
@@ -2,7 +2,7 @@
 using HarmonyLib;
 using TaleWorlds.CampaignSystem;
 
-namespace HaremKingv3.Patches
+namespace HaremKing.Patches
 {
     [HarmonyPatch(typeof(CharacterObject), "IsFemale", MethodType.Getter)]
     internal class HaremKingFemalePatch
@@ -12,14 +12,13 @@ namespace HaremKingv3.Patches
             Occupation occupation = __instance.Occupation;
             bool IsPlayerCharacter = __instance.IsPlayerCharacter;
 
-            List<Occupation> excludedOccupations = new List<Occupation>()
+            List<Occupation> femaleOccupations = new List<Occupation>()
             {
-                Occupation.NotAssigned,
-                Occupation.Lord,
-                Occupation.Special,
-                Occupation.NumberOfOccupations,
+                Occupation.Mercenary,
+                Occupation.Soldier,
+                Occupation.Bandit
             };
-            if (IsPlayerCharacter || excludedOccupations.Contains(occupation))
+            if (IsPlayerCharacter || !femaleOccupations.Contains(occupation))
             {
                 return true;
             }