+++ CHANGELOG.md
@@ -11,3 +11,21 @@ Add change log.
 Update Hamony library and mod xml for game version 1.2.9
 
 ### Fixed
+
+## [1.2.0]
+
+### Added
+
+**Companion limit patch**: increase companion limit to 3 times.
+
+**Disable donation patch**: disable garrison donation if the party belongs to player's clan.
+
+LICENSE and README.
+
+### Changed
+
+**Breaking Changes**: Rename the project from HaremKingv3 to HaremKing.
+
+The female patch now will only make soldier, mercenary and badit always female. Other occupations will not be changed.
+
+### Fixed